refactor: 체스에 알맞게 패키지 변경
- pawn  : com/wootecam/chess/pieces/
- board : com/wootecam/chess/
